Fixes https://github.com/linkerd/linkerd2/issues/2435