@hawkw 2019-03-10
@zaharidichev it looks like the CI build is currently failing because of the rustfmt formatting check. Can you run rustfmt on this branch and push the reformatted code?

As an aside, if you're using an editor that supports this, you might find it helpful to configure your editor to run rustfmt on save, at least for the `linkerd2-proxy` repository.