@olix0r 2019-03-11
It looks like CI is failing because `rustfmt` needs to be run in a few places.