@zaharidichev 2019-03-11
Sorry was a bit busy the last day. I will run it through rustfmt and try to collect some benchmarks as well as I am interested in the results too.