@zaharidichev 2019-03-12
@hawkw 

I ran the benchmarks that you suggested and here are the results I get

![bench](https://user-images.githubusercontent.com/4391506/54232466-64d6ba80-4513-11e9-9e00-558d6ae4ff68.jpg)


Here is the raw data: https://gist.github.com/zaharidichev/6407e3b8a9c0579b3a477b3596f10641

This is averaged across samples. It will be great to know whether the performance regression isexpected. Is it something that is acceptable and moreover is there adequate explanation for it ?