@hawkw 2019-03-12
@zaharidichev That's definitely interesting --- since you've averaged across multiple runs, I'd really like to see standard deviations for those measurements.