@zaharidichev 2019-03-12
@hawkw  I hope this helps

| BC                         | S0    | S1    | S2    | AVG   | STDEV |
|----------------------------|-------|-------|-------|-------|-------|
| Before H1 Request rate     | 992   | 992   | 995   | 993   | 1     | 
| Before H1 p50 latency (us) | 1844  | 1800  | 1789  | 1811  | 29    |
| Before H1 p75 latency (us) | 2473  | 2429  | 2413  | 2438  | 31    |
| Before H1 p90 latency (us) | 3255  | 3217  | 3181  | 3218  | 37    |
| Before H1 p95 latency (us) | 4137  | 4049  | 3942  | 4043  | 98    |
| Before H1 p99 latency (us) | 7467  | 7051  | 6785  | 7101  | 344   |
| Before H1 p99 latency (us) | 19304 | 16941 | 14810 | 17018 | 2248  |
| Before H2 Request rate     | 1000  | 1000  | 1000  | 1000  | 0     |
| Before H2 p50 latency (us) | 1604  | 1573  | 1514  | 1564  | 46    |
| Before H2 p75 latency (us) | 2209  | 2162  | 2048  | 2140  | 83    |
| Before H2 p90 latency (us) | 3084  | 3030  | 2878  | 2997  | 107   |
| Before H2 p95 latency (us) | 3865  | 3776  | 3484  | 3708  | 199   |
| Before H2 p99 latency (us) | 6220  | 6000  | 5765  | 5995  | 228   |
| Before H2 p99 latency (us) | 18954 | 17208 | 16144 | 17435 | 1419  |
| After H1 Request rate      | 998   | 998   | 1000  | 999   | 1     |
| After H1 p50 latency (us)  | 1998  | 1995  | 2221  | 2071  | 130   |
| After H1 p75 latency (us)  | 2831  | 2827  | 2920  | 2859  | 53    |
| After H1 p90 latency (us)  | 3812  | 3816  | 4050  | 3893  | 136   |
| After H1 p95 latency (us)  | 4396  | 4404  | 5082  | 4627  | 394   |
| After H1 p99 latency (us)  | 9114  | 9175  | 8014  | 8768  | 653   |
| After H1 p99 latency (us)  | 18803 | 18224 | 21225 | 19417 | 1592  |
| After H2 Request rate      | 1000  | 1000  | 1000  | 1000  | 0     |
| After H2 p50 latency (us)  | 2188  | 2193  | 2221  | 2201  | 18    |
| After H2 p75 latency (us)  | 2808  | 2832  | 2920  | 2853  | 59    |
| After H2 p90 latency (us)  | 3824  | 3887  | 4050  | 3920  | 117   |
| After H2 p95 latency (us)  | 4627  | 4786  | 5082  | 4832  | 231   |
| After H2 p99 latency (us)  | 7740  | 7806  | 8014  | 7853  | 143   |
| After H2 p99 latency (us)  | 18134 | 19091 | 21225 | 19483 | 1582  |