@carllerche 2019-03-12
As a general note, the `tokio-sync` mpsc channel's `buffer` argument means something different than `futures::sync::mpsc`. If the buffer size is small (less than the number of `Sender` handles), this could be the cause for the performance regression.

The old channel had a capacity of `buffer + num-sender-handles` where as the new channel has a buffer size of `buffer`.

However, off the top of my head, I am not sure which channels are being hit during the benchmark.