@olix0r 2019-03-26
@zaharidichev Sorry for the silence. We still want this! We'd like to do a bit more to diagnose the performance differences, but we've been focused on the upcoming 2.3 release recently.