@zaharidichev 2019-04-03
@carllerche Sorry I find it a bit hard to dissect. Where is exactly this pre - allocation happening. Do oyu think you can point me to this part of the code ?