@carllerche 2019-04-03
The channel is a [linked list of blocks](https://github.com/tokio-rs/tokio/blob/master/tokio-sync/src/mpsc/block.rs). Each block holds [32 messages](https://github.com/tokio-rs/tokio/blob/master/tokio-sync/src/mpsc/block.rs#L42).

The idea is that usually, when using a channel, many messages will be sent. By allocating storage for messages in blocks, allocations can be reduced.

You can kind of think of the channel as a lazy, infinite, array. When sending into the channel. First [an index for the message is acquired by incrementing a counter](https://github.com/tokio-rs/tokio/blob/master/tokio-sync/src/mpsc/list.rs#L62). Next, the [block for that index is found](https://github.com/tokio-rs/tokio/blob/master/tokio-sync/src/mpsc/list.rs#L65). If the block does not exist yet, it [is allocated](https://github.com/tokio-rs/tokio/blob/master/tokio-sync/src/mpsc/list.rs#L125). Finally, the message is written into [its index](https://github.com/tokio-rs/tokio/blob/master/tokio-sync/src/mpsc/list.rs#L69).

This provides significant gain when sending many messages, but regresses in some other cases.