@zaharidichev 2019-04-04
Thanks for the explanation @carllerche. That makes sense. So I ran the benchmarks again and modified the block cap size to be equal to 2 in one of my runs.  Here are the results of my runs. Do you think there is enough performance improvement in the version with BLOCK_CAP=2 to tell us that the preallocation logic is the root of the perf regression ? 

BEFORE: 
---------
    p50 latency (us): 1778
    p75 latency (us): 2561
    p90 latency (us): 3702
    p95 latency (us): 4499
    p99 latency (us): 7396
    p99 latency (us): 13383

AFTER (Block Cap 32):
-----------------------
    p50 latency (us): 2207
    p75 latency (us): 2851
    p90 latency (us): 3777
    p95 latency (us): 4422
    p99 latency (us): 8077
    p99 latency (us): 21796

AFTER (Block Cap 2):
-----------------------
    p50 latency (us): 1987
    p75 latency (us): 2597
    p90 latency (us): 3565
    p95 latency (us): 4345
    p99 latency (us): 7581
    p99 latency (us): 20620