@seanmonstar 2019-04-04
It *looks* to me like there may be something else, since the change from 32 to 2 was tiny (though noticeable).