@carllerche 2019-04-04
@zaharidichev another thing to try would be to switch all the channels to unbounded channels and see how that impacts the benchmark