@carllerche 2019-04-05
Hmm. Ok at this point, I guess profiling is in order.